provide consistent environement to test SSL libraries with 3rd party tool
the idea is to have a script (set of scripts) which build and install
environement such we can evaluate SSL performance of different
SSL implementations.

The initial version supports httpd from apache. The idea is to build
desired SSL library and install it to custom location (INSTALL ROOT).
Then we build apache and siege client and install it to INSTALL ROOT.

The proposed approach uses shell script as it is very lightweight
and portable between various *NIX flavours.

Reviewed-by: Neil Horman <nhorman@openssl.org>
(Merged from https://github.com/openssl/perftools/pull/49)