Add cmake "run" target and use it in PR CI
Introduce a run target, that executes all the test with different options.
For now is used only for minimal sanity checking, that are test can be run
and don't fail, but later can be extended to check the results for sanity
as well, and the design has some provisions put in place to be useful
for test data collection as well, for example.

Reviewed-by: Neil Horman <nhorman@openssl.org>
(Merged from https://github.com/openssl/perftools/pull/53)