.github/workflows/test.yml: limit ossl building parallelism
"make -j" enables to unlimited concurrency during execution
of the openssl build step, which sometimes leads to failures
due to resource exhaustion. Try to avoid that by limiting
the number of concurrently executed jobs to 4, similarly
to the way it is done elsewhere in the workflow steps
(including building openssl on FreeBSD).

Signed-off-by: Eugene Syromiatnikov <esyr@openssl.org>